Add GRPC service for load testing
Summary: We want to be able to load test GRPC services. This creates a GRPC load tester that should match what we are doing with the HTTP load tester. For the 4 types of GRPC endpoint (unary, client streaming, server streaming, and bidirectional streaming) we have an endpoint that takes sequence IDs and payloads/bytes requested, and the server will return the requested volume of data.

Test Plan: Ran it locally with bazel, need to more thoroughly test the streaming endpoints once I add the client.

Reviewers: zasgar, vihang, jamesbartlett, oazizi, #third_party_approvers

Reviewed By: zasgar, #third_party_approvers

JIRA Issues: PP-3238

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10695

GitOrigin-RevId: d0255aa9ba888da0cc31f893f1b14b0b789cf8b2