PerfProfiler: handle PEM restarts in Java symbolizer.
Summary: We affix both the pid of the Java process and its start time to the symbol file (both stable across PEM restarts). If the PEM restarts, we find the previous symbol file and *do not* try to re-attach the symbolization agent.

Test Plan: Need to write a test.

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Subscribers: yzhao

JIRA Issues: PP-2982

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10694

GitOrigin-RevId: 7e9b34b4df3041902a8544dac78c71d21d2584bf