Add parser for /proc/<pid>/status
Summary:
We are hoping to add debug UDTFs to report memory consumption stats
from our agents. This is just the parser for /proc/<pid>/status to parse the
necessary data. A future diff will add the UDTF to report this data.

Test Plan: Added a unit test.

Reviewers: zasgar, michelle, jamesbartlett, #stirling, oazizi

Reviewed By: #stirling, oazizi

Subscribers: oazizi

JIRA Issues: PP-3241

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10706

GitOrigin-RevId: 553adbf79c6d89014ccb1d266bb2934da5fd1802