Add UDTFs for _DebugAgentProcStatus and _DebugAgentProcSMaps
Summary:
We would like to dump /proc/<pid>/status and /proc/<pid>/smaps for
debugging purposes. We already have parsers for them, this adds the necessary
UTDFs for the same.

TODO: This seems to not work on Kelvin at the moment (I think it is due to the
getpid call). Fix it before submitting.

Test Plan:
Deploy dev vizier to cluster using skaffold. Run the new UDTFs and
ensure that the output matches what we see in the system.

Reviewers: zasgar, michelle, jamesbartlett

Reviewed By: jamesbartlett

JIRA Issues: PP-3241, PP-3242

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10711

GitOrigin-RevId: 8de54d7cae1ccbeb731c5b9e91ac2f41648fb117