Set a smaller store_limit for v2c.*.*.DurableMetadataUpdate
Summary:
The DurableMetadataUpdates are processed by VZMgr and converted into
MetadataIndex messaged which are also on STAN. We would like to provide a 24hr
retention for the latter to be consumed by the Indexer but I don't think we need
an equally long retention for the DurableMetadataUpdates since we have the
MissingMetadata(Request/Reponse) channels to grab any updates we might miss.

Test Plan: Deploy to dev, ensure that stan + cloud work fine.

Reviewers: michelle, philkuz, nserrino

Reviewed By: michelle

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10720

GitOrigin-RevId: 6ed33e19e51785ceeae8821e051831270177858d