Cleanup nats/etcd operator removal
Summary:
We had versioned the nats/etcd CRDs and types to assist us in removing
the operator controlled versions of each of them. At this point all consumers
should be using the stateful sets and we no longer need this code.

Test Plan: N/A removal of dead code. An operator RC should continue to work.

Reviewers: michelle, philkuz

Reviewed By: michelle

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10728

GitOrigin-RevId: c728f29d90ef693014a9465598c8e86b9b1fdfca