Fix pebble ttl deletion implementation
Summary:
The pebble ttl deleter was using strings.Split with `/` to check for
keys to delete. However our keys have `/` in them which makes the ttl deleter
check for entirely the wrong set of keys to delete.
The unit test had no keys with `/` in them and hence didn't exhibit the same
behavior.

This fixes the implementation and pulls out the ttl key formatters into helper
funcs to make it a bit more clear as to what is happening.

Test Plan:
Updated the TTL test to use a key with a `/` in it. The test should
exhibit the broken behavior without the corresponding changes to the TTL code.

Reviewers: michelle, nserrino

Reviewed By: michelle

JIRA Issues: PP-3249

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10734

GitOrigin-RevId: 700805dd7d446a1cb9ec62dea7891d19e2ea042b