ConnTracker: MemUsage API to report back approximate memory usage
Summary:
To support an eventual constrained memory mode, we need to know the approximate memory use of each ConnTracker.

This diff adds an API for getting the memory use in the data buffers and events of the ConnTracker.

Test Plan: Basic test included to check that memory use is reported back approrpiately. Because of the approximate nature of this function, there is some leeway in the test.

Reviewers: #stirling, jamesbartlett, yzhao

Reviewed By: #stirling, yzhao

Subscribers: yzhao

Signed-off-by: Omid Azizi <oazizi@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10746

GitOrigin-RevId: 99face5da7995d454c5230492d5e5323ef198320