Handle HTTP response with no content-length
Summary:
In Issue #370, a bug was reported that HTTP responses with non content-length were not handled correctly.
This diff adds a state to the HTTP parser and records whether the connection has closed in the state.
If the connection has closed, we take what's in the buffer as the response body, otherwise, we wait for
more data until the connection closes.
This fix addresses the above mentioned bug.

Test Plan: Updated http parse_test

Reviewers: #stirling, oazizi

Reviewed By: #stirling, oazizi

Subscribers: yzhao, oazizi

Signed-off-by: Ryan Cheng <rcheng@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10520

GitOrigin-RevId: dbcb50b144bba1b8ec588213654c8383af18f588