Add more error logging to operator deploys
Summary:
We're seeing cases where the operator is failing to deploy Vizier, but has been difficult to debug without more information.
This area of the code lacks good error logging, so that we can atleast know where in the deployment process things have started failing.

Test Plan: This is just adding logs and affects no functionality

Reviewers: vihang

Reviewed By: vihang

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10753

GitOrigin-RevId: 4f88a15a728c2017b3e41500c85c2c384cc0ea7a