Add new helpers to use jwx instead of jwt-go
Summary:
We are using a preview version of jwt-go and the released version
doesn't include all the features we need. So let's move to using jwx for all
our JOSE needs. This addes the helpers needed by the different consumers.

Test Plan: Added unit tests for the newly added helpers.

Reviewers: michelle, zasgar

Reviewed By: zasgar

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10755

GitOrigin-RevId: e8247944591ed68c7f55d54ebd25546c335acee5