Fix incorrect namespace in sock-shop demo yaml.
Summary: This yaml file for the Sock Shop demo was creating the wrong namespace. This bug probably went undetected because the CLI manually creates the namespace before deploying the yaml file when running `px demo deploy`. So deploys using the CLI were creating both the `px-sock-shop` and (unnecessary) `sock-shop` namespaces.

Test Plan: Built the CLI locally and pass it the dev artifacts.

Reviewers: michelle, vihang

Reviewed By: vihang

Signed-off-by: Hannah Troisi <htroisi@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10778

GitOrigin-RevId: 80b80b72630c4cabed41f9941950e8bddb41531d