[PP-3211] Add pgsql wire data generation.
Summary:
For the purposes of the socket tracer benchmark, we want to be able to generate a bunch of wire data for each protocol. This diff adds the utilities to make data generation easier for pgsql. These utilities could also be used in the tests instead of the ad-hoc char[] byte buffers that the tests currently used. I added a couple examples in the tests to demonstrate this usage.

This diff also restructures the build file, such that the test data is not compiled with the main cc library.

Test Plan: No functional changes, relying on existing tests. Note that this diff increases test coverage.

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Subscribers: yzhao

JIRA Issues: PP-3211

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10786

GitOrigin-RevId: 6d379b8aac9964bc27959f0378483d29d911409b