Don't validate JWT in CLI
Summary:
Since the CLI runs on a user's machine, there's a large chance of a
clock skew which would cause JWTs to not be valid. We used to not validate these
JWTs when parsing with jwt-go, but I added validation when we moved to using
jwx. Drop the validation so that CLI functions continue to work on machines with
large skew.

Test Plan: Pixie CLI should login properly on machines with a large skew.

Reviewers: michelle, zasgar

Reviewed By: michelle

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10783

GitOrigin-RevId: c42f675266fb41b878d6daa5b618af2116ea8675