Add semantic type rules for px.abs
Summary: I tried to do a px.abs on the result of px.parse_duration, but noticed that it didn't preserve the DURATION_NS type. This diff adds the correct semantic type to the result of px.abs.

Test Plan: not needed, we use these same rules elsewhere, just need to add them here

Reviewers: philkuz, jamesbartlett

Reviewed By: jamesbartlett

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10809

GitOrigin-RevId: d7a22eac106a04fa7ac9c806457b594f18b1887d