Remove opname from operator docs generator
Summary:
We used to use the opname for the docs. But when I was reading the docs, I found this to
be immensely confusing. I switched over to using the PxL name instead.

This change was used to make the output docs: https://github.com/pixie-io/docs.px.dev/pull/143

Test Plan: Rendered the output docs in local dev.

Reviewers: vihang, michelle, nserrino

Reviewed By: michelle

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10819

GitOrigin-RevId: 48674fa9b715b579459769fe09b033a89632e7bf