PerfProfiler: change args to java agent to indicate the symbol file path (vs. name stem, which was awkward).
Summary: Previously, the symbolization agent args indicated a file name stem because the symbol file name was uniquified based on target process pid. The name was given as a name stem (in case the agent needed to create a different file name for the txt log file). This setup was awkward both because it was a name stem and because it required coordination of the symbol file name in several places. Now we uniquify based on injecting the target process pid into the artifacts path. In this diff, we change the JVMTI symbolization agent args to indicate only the artifacts path and place a standard symbol file name in the shared header file.

Test Plan: Existing.

Reviewers: #stirling, oazizi

Reviewed By: #stirling, oazizi

Subscribers: oazizi

JIRA Issues: PP-2982

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10818

GitOrigin-RevId: 902ec31bb1f306ded30f5e9714801c306207506c