Memory leak fix in mmap event tracker structure
Summary:
A config flag controls whether we scan for mmap events, which is in turn used to detect dynamically loaded libraries.

When this flag is off, there could be a leak, since we add to the map but never drain it.

In this diff, we make sure not to add any items to the map if the flag has disabled the feature.

Test Plan: Rely on existing tests to ensure no functionality is broken. Manual check for the leak fix.

Reviewers: #stirling, jamesbartlett

Reviewed By: jamesbartlett

Signed-off-by: Omid Azizi <oazizi@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10830

GitOrigin-RevId: b5bd65214fbe9d7f17f7ea1edb4cd779828af022