Add px._HeapReleaseFreeMemory UDTF.
Summary:
Adds a UDTF that calls ReleaseFreeMemory on all the agents.
This causes tcmalloc to release spans in the free list to the OS. This is useful for experimenting with memory usage.

Test Plan: Checked that px._HeapStats, showed about 600MB of free spans in the page heap, then after running the UDTF, another call to `px._HeapStats` showed that 500MB of that was returned to the OS.

Reviewers: oazizi, vihang, nserrino

Reviewed By: vihang

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10846

GitOrigin-RevId: 7acee822d2514fa562ea1895ab7bcee54e1b5f56