PC-1437 Add database tables for data retention plugin config
Summary:
This diff adds the database tables which will be used to track a plugin's configuration.
`data_retention_plugin_releases` will track the provider-specified configuration options for their plugin. For example, the preset scripts and config values that need to be filled out when enabling the plugin. This will be populated by a job, similar to artifact tracker.
`org_data_retention_plugins` tracks an org's settings for a plugin, for example: their plugin version, what their values for each config setting are.

This does not include a table for the longterm retention scripts that a user has yet. That will follow in a separate diff.

Test Plan: Skaffold deploy, ensure that the migrations run. These tables are currently unused.

Reviewers: vihang, nserrino

Reviewed By: vihang

JIRA Issues: PC-1437

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10855

GitOrigin-RevId: d47f1bdaf1ed30f0511bc8fbaf7e5b753ebafcd0