Add script for getting the correct binaries for symbolizing heap profiles.
Summary: Adds a script (`scripts/download_heap_prof_mapped_files.sh`) that takes as input a heap profile in pprof format, and the node the heap profile was collected on. Then it sshes into the given node, collects all the files that correspond to mapped files in the heap profile, and ships them back to your local machine as a tar archive. The archive is then extracted into a tmp folder that can subsequently be used as the `PPROF_BINARY_PATH`. The script is designed to work even if the workload that was profiled is running in a container, as it looks for files relative to `/proc/$PID/root` where `PID` is found based on mapping entries in the heap profile and the corresponding running processes on the given node.

Test Plan: Tested that running the script dumps the expected files into a temp directory. Also tested that the files output, yield valid heap profiles when used for symbolization in pprof.

Reviewers: #engineering, zasgar

Reviewed By: #engineering, zasgar

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10867

GitOrigin-RevId: 37a75abf2624829c3808fc5fded8965973504883