Split socket_trace_connector_test
Summary:
SocketTraceConnectorTest is a large test that has some protocol specific aspects to it.

This diff removes the protocol specific parts to a different file.
This also sets the stage to eventually removing those protocol specific parts,
since that coverage really belongs in the protocol parser testing.

Test Plan: This is test updates.

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Signed-off-by: Omid Azizi <oazizi@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10876

GitOrigin-RevId: ff181de9e13ad82b8e272e44df08da37172fc590