Replace dataframe tests with ParseScript() to simplify QLObjects calls
Summary:
Part of a way to improve the readability of tests in QLObjects.
The existing objects approach is arduous, repetitive and hard to read. You have to create the
objects, extract the exact method out, create the argument as a function.

It's basically impossible to understand what's being tested.

This diff proposes a new way: using actual mini scripts. By pushing up the var_table argument
from ast_visitor, we can actually use the ast_visitor in the middle to do the arduous function
creation, object creation, etc.

I demonstrate the value of this by rewriting  the Dataframe tests. Now future readers
can just read the PxL code, which they'll be familiar with, and understand what's going on in a
test.

Test Plan: Updated a test and made sure that it works.

Reviewers: jamesbartlett, nserrino

Reviewed By: jamesbartlett

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10889

GitOrigin-RevId: 1fa487374508e2d7774ef5ced010660b0391e46e