PP-3268: Integrate feature flag generation into Pixie Cloud config_manager service
Summary: Depends on D10881. This diff hooks up the feature flag generation added in D10881 (and uses the internal endpoint added in D10880). After this diff lands, the infrastructure will be in place to add feature flags from a feature flag provider such as LaunchDarkly to Vizier. If LaunchDarkly is not configured, then the provider will always select the default value for the flag. This will allow us to use feature flags in Vizier deployments, not just the UI.

Test Plan: End to end test with new Vizier deployment

Reviewers: michelle, vihang, zasgar

Reviewed By: michelle

JIRA Issues: PP-3268

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10884

GitOrigin-RevId: d16eb2cc04503e7239e3c43386b8f8610c3a118b