PP-3268: Support config manager looking up org ID from deploy key
Summary: We want Pixie cloud's Vizier CRD generation to include feature flags. In order to know which feature flags to use for a given Vizier deployment, the config manager needs to use the org ID for that Vizier and look up the value for that org in the feature flag provider (such as LaunchDarkly). The config manager has a deployment key and that deployment key can be used by the config manager to look up the org ID. This diff ensures that auth for the public deploy key manager endpoints is performed in the API server (like our other services). It makes it possible for the config manager, an internal service, to look up the org ID for a given deploy key.

Test Plan: Added unit tests, also ran an end to end test with this + upcoming stack of diffs

Reviewers: michelle, vihang, zasgar

Reviewed By: michelle

JIRA Issues: PP-3268

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10880

GitOrigin-RevId: 40a2b780d51af4280a8c1307fc5017f5284893f7