Expose the expr directly on an ExprObject
Summary:
I am creating a series of changes where I switch over from storing the IRNode on all
QLObjects and instead storing an Operator in the Dataframe objects and ExpressionIR in ExprObjects.

This pattern was very unclean because there are cases where we don't even want an IRNode associated
with an object. That wasn't clear in the initial design but is now more obvious with these more
temporary object types.

Test Plan: Tested in later diffs by switching references from node() to expr()

Reviewers: jamesbartlett, vihang, michelle

Reviewed By: vihang

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10921

GitOrigin-RevId: 51809b36ce4f1d9a5bfe6bc719e64de9c1a91b41