Simplify QLObject tests using ParseExpression()
Summary:
In an earlier diff I added support for ParseExpression. This diff now updates
the remaining tests to use this expression. It makes testing much more readable and easier as well.

Test Plan:
Ran the updated tests and they pass. The test verify that certain fields are set which
apparently is the case here still.

Reviewers: jamesbartlett, vihang, michelle

Reviewed By: jamesbartlett

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10922

GitOrigin-RevId: 174337b14da9231acdae8d22c5ccbd9d3c2d5b82