[PC-1452]: Smarter retries for mutation scripts
Summary:
Retry logic for mutations scripts worked like this:
- Start running script
- Wait five seconds
- If not done and script has mutations, restart script (and leave results as-is)

If this timer expired when results were already coming in, the UI would crash because results for one query were trying to overwrite results for a different query (intentional crash because this situation means something broke).

What it should do (and does with this diff):
- Start running script
- Start a five second timer if script has mutations
- If timer expires and probes aren't deployed yet, retry
- Clear timer if probes finish deploying, even if script isn't done running yet

Test Plan:
Try running a mutations script that deploys its probes fairly quickly, but takes a while to run.

As of this writing, a good example is `pxbeta/vfs_snoop`.

Running it for the first time on a cluster should:
- Show the mutation modal
- Eventually put a check mark on the probe
- Eventually show results if not cancelled

Running it on a cluster that already has that probe deployed should:
- Not show the mutation modal
- Wait like any other script for results to come in

In both cases, cancelling the script execution should work normally.

Reviewers: michelle, vihang

Reviewed By: michelle

Subscribers: oazizi

JIRA Issues: PC-1452

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10919

GitOrigin-RevId: 69f2f5a6730037c9e47a9683eec877c8b21784e8