PerfProfiler: add a test case for JavaSymbolizer.
Summary: We add a test case for JavaSymbolizer by creating a fake java binary that writes its own symbol file. The fake java binary is detected as a Java application triggering the java symbolization logic. We need this fake java binary such that we have a crafted symbol file (because with an actual JVM, we cannot predict the in-memory location of the generated symbols.

Test Plan: This is a test.

Reviewers: #stirling, oazizi

Reviewed By: #stirling, oazizi

Subscribers: oazizi

JIRA Issues: PP-2982, PP-3265

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10932

GitOrigin-RevId: d252757f3b78d24aba7449e2f32faefd4c7e663c