Default theme var when embed doesn't specify
Summary:
We were already doing this, but incorrectly.

Uses `deepmerge` from `@mui/utils`, which has to be specified as a dependency. Our existing MUI dependencies already included it, but didn't expose it for direct import.

Test Plan:
Use the embed instructions to load in an iframe; run px/perf_flamegraph.
Graph should render instead of erroring on a missing `baseRGB` variable.

Reviewers: michelle, vihang, #third_party_approvers

Reviewed By: michelle, #third_party_approvers

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10949

GitOrigin-RevId: 69b61525abd6074702bab6ac8485e5442c8a4914