Add PDB to stan and update the stan cluster size
Summary:
The PDB ensures that stan never loses quorum. Update the cluster size
to 3 isntead of 2 because, 3 would allow for 1 disruption, 2 is equivalent to 1
from a raft disruption perspective.

Test Plan:
Deployed to my cluster, went through a nodepool upgrade. Stan didn't
lose quorom as long as the nodepool had surge set to 1.

Reviewers: zasgar, michelle

Reviewed By: michelle

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10972

GitOrigin-RevId: 60eee22c6e73590e6126b987cfbe75c539905bd4