PP-3298: Remove deprecated CLI flag to set table store size
Summary:
We have a deploy flag, --table_store_table_size in the px deploy flow. However, it doesn't actually work or do what is expected (which is cap the table store size). Removing it, and in the next set of diffs will add a replacement that does what is expected.

After this lands and the CLI is released, I will remove the piping from the Vizier CRD that contains this flag. and hopefully remove the environment variable that it sets as well.

Test Plan: build CLI

Reviewers: michelle, jamesbartlett, zasgar, oazizi

Reviewed By: oazizi

JIRA Issues: PP-3298

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10985

GitOrigin-RevId: 173cd06bfdb08ab39af839394ef03028b22e63d9