PP-3301: Add min_time to px._DebugTableInfo() UDTF
Summary: We want to be able to create a PxL script that will help users figure out how much memory to allocate to the table store in order to get their desired level of retention. This diff adds a `min_time` column to that UDTF (which is also useful in general) in order to support that PxL script.

Test Plan: ran the script

Reviewers: jamesbartlett, oazizi, zasgar

Reviewed By: jamesbartlett

JIRA Issues: PP-3301

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10991

GitOrigin-RevId: f13286388560f09082c98c9c78c8c14a7c7ab9c5