PP-3275: Update proto with column changes and scoping down
Summary:
Proto was using column names instead of indices which is standard for the plan proto
api. I've fixed that.

I missed the resource configuration as part of this, so added that in here.

Finally: I decided to scope down some of the features we're supporting to make this simple.

Test Plan: n/a no code changes.

Reviewers: michelle, jamesbartlett, vihang

Reviewed By: michelle

JIRA Issues: PP-3275

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10971

GitOrigin-RevId: 220271692dfae092c6a1fd736e63a1dbefecdce6