UI: Fix issue where non-latency columns are colored like latency columns
Summary:
We have special coloring for latency column duration values, so users can understand how long a request took or service latency statistics more easily. However, not all durations are latencies, and so they shouldn't be colored red just because they are a duration. We currently do this check (if it's a latency column) properly with duration quantiles, but some wires must have gotten crossed in a refactor when it comes to pure durations. This diff corrects the issue and makes it consistent between quantiles and non-quantiles.

This was identified when working on PP-3302, since it affects some views I want to create.

Test Plan: unit, ran ui

Reviewers: nlanam, michelle

Reviewed By: nlanam

JIRA Issues: PP-3302

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10999

GitOrigin-RevId: 964c1bccd9991a116042e8e2008788df45980668