PP-3302: Add more units to duration formatting in the UI
Summary: We want to use the duration type to visualize things like pod age, and data retention period. However, the current units for a human readable duration only go up to seconds. This makes it very difficult when reading a duration in days. In this diff, (as well as D10997 for the CLI), days, minutes, and hours are added as an option. The common case of latency of 60 seconds or less will remain the same.

Test Plan: ran the UI , unit tests, tried existing

Reviewers: nlanam, michelle, jamesbartlett, zasgar

Reviewed By: nlanam

JIRA Issues: PP-3302

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11006

GitOrigin-RevId: ed955a98fc0a34b4173aff9fdbd8e9ffaa7b2a24