Move std::visit helper into the common/utils.h file
Summary:
std::variant is a nice way to combine multiple types together. The recommended usage when
disambiguating the types is to use std::visit. In c++20 you can pass a list of lambdas into the
arguments. Before we support c++20, you would need a complex set of lambdas to make this work.

There is one solution, suggested in cppreference to create a helper type, that makes this super
clean: https://en.cppreference.com/w/cpp/utility/variant/visit

I've added that change into the common/utils.h.

This is already in use in otel implementations but I thought this would be generally useful, hence
adding to the common repo.

Test Plan: code already in use, just making it more available

Reviewers: jamesbartlett, oazizi, vihang, zasgar

Reviewed By: jamesbartlett

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11031

GitOrigin-RevId: 56aa0de420aa319436bfb0a5e8e22010b99d4e2a