[DataStreamBuffer] Clear AlwaysContiguousDataStreamBuffer when there are no more valid chunks.
Summary: Previously, the AlwaysContiguousDataStreamBuffer would only be considered empty if the underlying buffer was empty and not if there were no valid chunks. In theory, this should be fine, but we are running into an unexpected scenario where there are no valid chunks but the underlying buffer is not empty. Regardless of the underlying cause of this scenario, this diff makes it so that whenever we have no valid chunks, we always clear the buffer_.

Test Plan: Used gdb to diagnose looping issue. Found that it occurs when there are no valid chunks but the underlying data buffer is non-empty. Also, skaffold deployed to test that things work as usual.

Reviewers: #stirling, oazizi

Reviewed By: #stirling, oazizi

Subscribers: oazizi

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11045

GitOrigin-RevId: 8ca9c838cd5a18f09c7cafb26231cbb51021b30e