PC-1462 Querybroker should sync cronScript state with cloud
Summary:
Queybroker is responsible for executing cronscripts. To do so, it needs to know the correct set of cronscripts it should be running.
Querybroker should keep its cronscripts synced with Pixie Cloud. to do so, we use the script checksum at startup to determine whether or not the full set of scripts should be fetched from the cloud.
This diff just covers the initial sync. The querybroker must also track future changes (new scripts, updates, deletions) by listening to the message bus.

Test Plan: unit tests

Reviewers: vihang, nserrino, philkuz

Reviewed By: vihang

JIRA Issues: PC-1462

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11047

GitOrigin-RevId: 336209ce731551a189d321c7fb63537f52a7cdc9