PC-1462 Update querybroker script runner to execute cronscripts
Summary:
This diff updates the script runner to actually execute the given cronscript at the provided frequency.
Currently, it does not do any error checking. However, in the future, we would likely want to record these statuses to surface to the user.

Test Plan: ran in skaffold, with a script stubbed in. ensure that it runs successfully at a given frequency.

Reviewers: vihang, nserrino, philkuz

Reviewed By: vihang

JIRA Issues: PC-1462

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11054

GitOrigin-RevId: 51e4942af000a491e7f33bec4cb12ab5e4aab7b2