PC-1464 Send cronscript deletion updates to Viziers
Summary:
As a part of sycning Vizier + Cloud cronscripts, the cloud is responsible for sending updates to Viziers whenever a cron script is created/updated/deleted.
This diff updates the cronscript service so that it sends out deletion updates whenever a cronscript is deleted. As NATS messages do not guarantee delivery, there is retry logic in hopes to ensure that the vizier does receive the message. However, retries may lead to Viziers getting messages out-of-order. Added a timestamp field so that Viziers can determine when to discard an update for a script. For example, if a vizier receives "Update B @ timestamp 2", and shortly after "Update A @ timestamp 1", it knows that it should ignore update A since it already has the latest changes.

Test Plan: unit test

Reviewers: vihang, nserrino, philkuz

Reviewed By: vihang

JIRA Issues: PC-1464

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11069

GitOrigin-RevId: 85290bfe3f7daa8a750c529f831dfd4991c0f772