PC-1464 Handle GetCronScriptsChecksum requests in cronscript svc
Summary:
As Viziers start up, they will try to sync with the cronscript service in the cloud to ensure they have the latest scripts. The first request they will make is
for the checksum of the scripts. If the checksum of the Vizier's script matches that of the cloud, they do not need to request all scripts.
This diff implements the handler for GetCronScriptsChecksum requests in the cronscript service. These messages are all sent across NATS using cvmsgs.

Test Plan: unit tests, deploy to testing cloud + skaffold deploy Vizier with script syncing enabled. Verify that messages are sent back and forth between Vizier + Cloud

Reviewers: vihang, nserrino, philkuz

Reviewed By: vihang

JIRA Issues: PC-1464

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11067

GitOrigin-RevId: 9fa6394eb3588760f6c2a830bfd28fa7681bda7a