Plugin svc: handle config updates that send "enabled" when already enabled
Summary:
before we were assuming that "enabled" would only be set when the plugin is indeed enabled when it was previously disabled.
we should just be more robust and check whether the plugin is actually getting disabled/enabled, rather than assuming the request is correct.

Test Plan: Deploy to skaffold

Reviewers: vihang

Reviewed By: vihang

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11108

GitOrigin-RevId: 4a627acbbb8b0301bf0779f35584b2bccef7d43b