PP-3326: Return VizierName in vzmgr.VizierConnected
Summary:
I want to get the VizierName into clusters that already have registered. I decided that
this data will be acquired during the "doRegistrationHandshake()" part of the vz-cloud bridge
setup on the vizier side. VizierConnected generates that data. Change was a simple addition of a
field

Alternatives considered:
1. Call the RegisterVizierDeployment in vizier if the VizierName wasn't set for the cluster. That
doesn't work because the deployment-key used during initial registration is no longer valid.
2. Create a new RPC for GetVizierName. The pipeline was a bit complicated, required a new rpc call
and a version had a bit of a hacky authentication scheme with JWT. This solution is more simple.

Test Plan: Updated the vzmgr test. Going to test e2e with the vizier side.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PP-3326

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11133

GitOrigin-RevId: a4af2432d7deec721902028eed9019c9f49b8685