Add TypeErasedArrowBuilder for type-erased use of arrow ArrayBuilders.
Summary:
Adds a TypeErasedArrowBuilder helper class, that wraps an arrow ArrayBuilder, such that the ArrayBuilder can be used uniformly regardless of the Array data type.
This will be used in the new table store diffs to come, to help clean up the array building code.

Test Plan: This diff doesn't affect existing functionality. This code is used in a future diff, and tested there.

Reviewers: nserrino, philkuz, vihang, zasgar

Reviewed By: nserrino

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11140

GitOrigin-RevId: 3286e9c5de06b2da83715e47cc71b97d5db40d4e