[table/internal] Add ArrowArrayCompactor.
Summary: Another utility that will be used in the coming refactor. This class allows building a compacted arrow Array for each column, from a set of slices of RecordOrRowBatch's.

Test Plan: Currently unused by production code, but added a test for this component, and also tested it integrated with other components in later diffs.

Reviewers: nserrino, philkuz, vihang, zasgar

Reviewed By: zasgar

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11144

GitOrigin-RevId: cd9e482a56cb556bf34bf391debfa884c8b9af70