[table/internal] Add BatchSizeAccountant.
Summary: Adds another table utility. to keep track of batch sizes (in bytes) for each batch (hot or cold). This information is then used at compaction time to determine what slices of the current hot batches to compact into a compacted cold batch.

Test Plan: Added test for this component. Component currently unused in production code. Tested in future diff that integrates it with the table.

Reviewers: nserrino, philkuz, vihang, zasgar

Reviewed By: nserrino

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11176

GitOrigin-RevId: 53dfa77e69ec02f61ea5cf92a3eb560dbb7357c4