[table/internal] Add StoreWithRowTimeAccounting.
Summary: Adds a wrapper around a std::deque that keeps track of time and unique row ID's for each batch in the store. Templated by `StoreType` which allows the very similar logic between hot/cold stores to be deduplicated, and only the batch access logic to be duplicated.

Test Plan: No changes to production code in this diff. But this new wrapper has a new test added, and it was tested integrated as part of a future diff.

Reviewers: nserrino, philkuz, vihang, zasgar

Reviewed By: nserrino

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11195

GitOrigin-RevId: a758b3b75cca510c1f7c0ef09228d3eedd9f8850